+++ common/buildcraft/lib/gui/ISimpleDrawable.java
@@ -10,7 +10,7 @@ import buildcraft.lib.gui.pos.IGuiPosition;
 
 @FunctionalInterface
 public interface ISimpleDrawable {
-    void drawAt(int x, int y);
+    void drawAt(double x, double y);
 
     default void drawAt(IGuiPosition element) {
         drawAt(element.getX(), element.getY());
