+++ common/buildcraft/lib/gui/help/DummyHelpElement.java
@@ -23,22 +23,22 @@ public class DummyHelpElement implements IGuiElement {
     }
 
     @Override
-    public int getX() {
+    public double getX() {
         return area.getX();
     }
 
     @Override
-    public int getY() {
+    public double getY() {
         return area.getY();
     }
 
     @Override
-    public int getWidth() {
+    public double getWidth() {
         return area.getWidth();
     }
 
     @Override
-    public int getHeight() {
+    public double getHeight() {
         return area.getHeight();
     }
